+++ packages/terra-button-group/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Removed unnecessary `wdio.conf.js` file
+  * Upgrade tests to terra-functional-testing and WebDriverIO v6
 
 ## 3.53.0 - (February 2, 2021)